Separating out Workbench as its own project/package
