+++ Source/Clients/AspNetCore/AspNetCore.csproj
@@ -2,7 +2,6 @@
     <PropertyGroup>
         <AssemblyName>Cratis.AspNetCore</AssemblyName>
         <RootNamespace>Cratis.AspNetCore</RootNamespace>
-        <GenerateEmbeddedFilesManifest>true</GenerateEmbeddedFilesManifest>
     </PropertyGroup>
 
     <ItemGroup>
@@ -10,32 +9,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="Newtonsoft.Json.Schema" Version="$(NewtonsoftJsonSchema)" />
-    </ItemGroup>
-
-    <ItemGroup>
         <ProjectReference Include="../DotNET/DotNET.csproj" />
     </ItemGroup>
-
-    <ItemGroup>
-        <EmbeddedResource Include="$(MSBuildThisFileDirectory)/workbench/**/*" />
-    </ItemGroup>
-    <!--
-    <Target Name="BuildWorkbench" Condition="$(Configuration) == 'Release'" BeforeTargets="BeforeBuild">
-        <PropertyGroup>
-            <WorkbenchDirectory>$(MSBuildThisFileDirectory)../../Workbench/Events/Web</WorkbenchDirectory>
-        </PropertyGroup>
-
-        <ItemGroup>
-            <FilesToCopy Include="$(WorkbenchDirectory)/wwwroot/**/*" />
-        </ItemGroup>
-
-        <Exec WorkingDirectory="$(WorkbenchDirectory)" Command="base_path=/events/ yarn build" />
-        <Copy SourceFiles="@(FilesToCopy)" DestinationFolder="$(MSBuildThisFileDirectory)/workbench" />
-
-        <CreateItem Include="$(MSBuildThisFileDirectory)/workbench/**/*">
-            <Output TaskParameter="Include" ItemName="EmbeddedResource"/>
-        </CreateItem>
-    </Target>
-    -->
 </Project>
