+++ cratis.sln
@@ -61,6 +61,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orleans", "Source\Extension
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Autofac", "Source\Extensions\Autofac\Autofac.csproj", "{C674775B-268B-4EFB-AAB3-ACFF4EE7063B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCore", "Source\Clients\AspNetCore.Workbench\AspNetCore.csproj", "{C44DD087-3074-4927-9D10-FDC2E76F6724}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -338,6 +340,18 @@ Global
 		{C674775B-268B-4EFB-AAB3-ACFF4EE7063B}.Release|x64.Build.0 = Release|Any CPU
 		{C674775B-268B-4EFB-AAB3-ACFF4EE7063B}.Release|x86.ActiveCfg = Release|Any CPU
 		{C674775B-268B-4EFB-AAB3-ACFF4EE7063B}.Release|x86.Build.0 = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|x64.Build.0 = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Debug|x86.Build.0 = Debug|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|x64.ActiveCfg = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|x64.Build.0 = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|x86.ActiveCfg = Release|Any CPU
+		{C44DD087-3074-4927-9D10-FDC2E76F6724}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{91140C30-869A-4A6F-9ED6-0E3D1F5A18E5} = {5D74C0B9-D937-4B72-A30F-EAB428577AC0}
@@ -364,5 +378,6 @@ Global
 		{6E0BFEA9-E647-40B3-BAB0-FC80971914C1} = {FA1BA5A7-A7BD-4A29-B8EF-84605EE09127}
 		{03D1706A-A9F4-4134-A131-220ECC37A9F7} = {FA1BA5A7-A7BD-4A29-B8EF-84605EE09127}
 		{C674775B-268B-4EFB-AAB3-ACFF4EE7063B} = {FA1BA5A7-A7BD-4A29-B8EF-84605EE09127}
+		{C44DD087-3074-4927-9D10-FDC2E76F6724} = {D2BFE0F8-49AE-4F63-9A26-5D0DDDA933C4}
 	EndGlobalSection
 EndGlobal
